LayoutTests/fast/dom/document-set-body-null-expected.txt

Tests setting document.body to null

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDocument.body = null threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS successfullyParsed is true

TEST COMPLETE
